test/functional-portable/discovery/testdata/python-flask/requirements.txt
flask>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
